Adicionando & logica estilização para ViewProjectModal
